+++ Source/CIM.RemoteManager.App/CIM.RemoteManager.Core/ViewModels/SensorDetailsViewModel.cs
@@ -310,8 +310,8 @@ namespace CIM.RemoteManager.Core.ViewModels
         /// <summary>
         /// Converting Unix to Windows DateTime
         /// </summary>
-        private double _sinceTimeStamp;
-        public double TimeStamp
+        private int _sinceTimeStamp;
+        public int TimeStamp
         {
             get => _sinceTimeStamp;
             set => SetProperty(ref _sinceTimeStamp, value);
@@ -346,8 +346,8 @@ namespace CIM.RemoteManager.Core.ViewModels
         /// <summary>
         /// Converting Unix to Windows DateTime
         /// </summary>
-        private double _minimumOccuranceTimeStamp;
-        public double MinimumOccuranceTimeStamp
+        private int _minimumOccuranceTimeStamp;
+        public int MinimumOccuranceTimeStamp
         {
             get => _minimumOccuranceTimeStamp;
             set => SetProperty(ref _minimumOccuranceTimeStamp, value);
@@ -381,8 +381,8 @@ namespace CIM.RemoteManager.Core.ViewModels
         /// <summary>
         /// Converting Unix to Windows DateTime
         /// </summary>
-        private double _maximumOccuranceTimeStamp;
-        public double MaximumOccuranceTimeStamp
+        private int _maximumOccuranceTimeStamp;
+        public int MaximumOccuranceTimeStamp
         {
             get => _maximumOccuranceTimeStamp;
             set => SetProperty(ref _maximumOccuranceTimeStamp, value);
@@ -825,7 +825,7 @@ namespace CIM.RemoteManager.Core.ViewModels
                             if (plotTime)
                             {
                                 // Plot time
-                                sensorPlot.UnixTimeStamp = splitSensorValues[i].SafeConvert<double>(0);
+                                sensorPlot.UnixTimeStamp = splitSensorValues[i].SafeConvert<int>(0);
                                 sensorPlot.TimeStamp = sensorPlot.UnixTimeStamp.UnixTimeStampToDateTime();
                                 plotTime = false;
